AMP-5080
git-svn-id: file:///svn/trunk/amp@8454 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ WEB-INF/src/org/digijava/module/aim/auth/AmpAuthenticationFilter.java
@@ -331,7 +331,7 @@ public class AmpAuthenticationFilter
             appSettings.setAppSettingsId(ampAppSettings
                                          .getAmpAppSettingsId());
             appSettings.setDefRecsPerPage(ampAppSettings
-                                          .getDefaultRecordsPerPage().intValue());
+                                          .getDefaultRecordsPerPage());
             appSettings.setCurrencyId(ampAppSettings.getCurrency()
                                       .getAmpCurrencyId());
             appSettings.setFisCalId(ampAppSettings.getFiscalCalendar()
